rename TransM -> Trans and remove old Trans
