+++ connector/src/main/scala/quasar/qscript/Trans.scala
@@ -17,62 +17,64 @@
 package quasar.qscript
 
 import slamdata.Predef._
+import quasar.contrib.matryoshka.transHyloM
 import quasar.fp._
 
 import matryoshka._
 import matryoshka.data._
-import matryoshka.implicits._
 import matryoshka.patterns._
-import scalaz._
+import scalaz._, Scalaz._
 
-trait Trans[F[_]] {
+trait Trans[F[_], M[_]] {
   def trans[T, G[_]: Functor]
       (GtoF: PrismNT[G, F])
       (implicit TC: Corecursive.Aux[T, G], TR: Recursive.Aux[T, G])
-      : F[T] => G[T]
+      : F[T] => M[G[T]]
 }
 
 object Trans {
-  def apply[T[_[_]]: BirecursiveT, F[_], G[_]: Functor]
-      (trans: Trans[F], t: T[G])
+
+  def apply[T[_[_]]: BirecursiveT, F[_], G[_]: Traverse, M[_]: Monad]
+      (trans: Trans[F, M], t: T[G])
       (implicit FG: F :<: G, FT: Injectable.Aux[G, QScriptTotal[T, ?]], B: Branches[T, G])
-      : T[G] =
+      : M[T[G]] =
     applyTrans(trans, PrismNT.inject)(t)
 
-  def applyTrans[T[_[_]]: BirecursiveT, F[_], G[_]: Functor]
-      (trans: Trans[F], GtoF: PrismNT[G, F])
+  @SuppressWarnings(Array("org.wartremover.warts.NonUnitStatements"))
+  def applyTrans[T[_[_]]: BirecursiveT, F[_], G[_]: Traverse, M[_]: Monad]
+      (trans: Trans[F, M], GtoF: PrismNT[G, F])
       (t: T[G])
       (implicit G: Injectable.Aux[G, QScriptTotal[T, ?]], BR: Branches[T, G])
-      : T[G] = {
+      : M[T[G]] = {
 
-    val transG: G[T[G]] => G[T[G]] =
-      gtg => GtoF.get(gtg).fold(gtg)(trans.trans[T[G], G](GtoF))
+    val transG: G[T[G]] => M[G[T[G]]] =
+      gtg => GtoF.get(gtg).fold(gtg.point[M])(trans.trans[T[G], G](GtoF))
 
     val prismT: PrismNT[QScriptTotal[T, ?], F] =
       PrismNT(
         λ[QScriptTotal[T, ?] ~> (Option ∘ F)#λ](x => G.project(x).flatMap(GtoF.get.apply)),
         G.inject compose GtoF.reverseGet)
 
-    def tb[A]: G[A] => G[A] = BR.branches.modify(transBranches[T, F](trans, prismT))
-    t.transHylo(tb[T[G]], transG)
+    def tb[A]: G[A] => M[G[A]] = BR.branches.modifyF[M](transBranches[T, F, M](trans, prismT))
+    transHyloM(t)(tb[T[G]], transG)
   }
 
   ////
 
-  @SuppressWarnings(Array("org.wartremover.warts.Recursion"))
-  private def transBranches[T[_[_]]: BirecursiveT, F[_]]
-      (trans: Trans[F], QTtoF: PrismNT[QScriptTotal[T, ?], F])
+  @SuppressWarnings(Array("org.wartremover.warts.Recursion", "org.wartremover.warts.NonUnitStatements"))
+  private def transBranches[T[_[_]]: BirecursiveT, F[_], M[_]: Monad]
+      (trans: Trans[F, M], QTtoF: PrismNT[QScriptTotal[T, ?], F])
       (fqs: FreeQS[T])
-      : FreeQS[T] = {
+      : M[FreeQS[T]] = {
 
     type G[A] = CoEnv[Hole, QScriptTotal[T, ?], A]
 
     val p: PrismNT[G, F] = QTtoF.compose(PrismNT.coEnv)
 
-    val transT: G[FreeQS[T]] => G[FreeQS[T]] =
-      co => p.get(co).fold(co)(trans.trans[FreeQS[T], G](p))
+    val transT: G[FreeQS[T]] => M[G[FreeQS[T]]] =
+      co => p.get(co).fold(co.point[M])(trans.trans[FreeQS[T], G](p))
 
-    def tb[A]: G[A] => G[A] = Branches.coEnv[T, Hole, QScriptTotal[T, ?]].branches.modify(transBranches[T, F](trans, QTtoF))
-    fqs.transHylo(tb[FreeQS[T]], transT)
+    def tb[A]: G[A] => M[G[A]] = Branches.coEnv[T, Hole, QScriptTotal[T, ?]].branches.modifyF(transBranches[T, F, M](trans, QTtoF))
+    transHyloM(fqs)(tb[FreeQS[T]], transT)
   }
 }
